Reduced  variable data type.

+++ 05-functions/solutions/08-SwapEveryTwoBits.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 
 
-int swapBits(int n, int i, int j) {
+int swapBits(int n, short i, short j) {
     if (((n >> i) & 1) != ((n >> j) & 1)) {
         n ^= (1 << i) | (1 << j);
     }
@@ -10,7 +10,7 @@ int swapBits(int n, int i, int j) {
 
 
 int swapEveryTwoBits(int n) {
-    for (int i = 0; i < sizeof(n) * 8 - 1; i += 2) {
+    for (short i = 0; i < sizeof(n) * 8 - 1; i += 2) {
         n = swapBits(n, i, i + 1);
     }
     return n;
